requirements.txt
discord.py-self
python-telegram-bot
re
BeautifulSoup
urlopen
Request